Resolved comments/reply delete, login user like/unlike issue in feeds,experience, Added isPrivate property in channel

+++ vire-model/src/main/java/com/vire/model/response/ExperienceDetailResponse.java
@@ -24,7 +24,7 @@ public class ExperienceDetailResponse {
     private MasterResponse categoryResponse;
     private Integer commentsCount;
     private Boolean loginUserLiked;
-    private Integer viewsCount;
+    private Long viewsCount;
     private Long createdTime;
     private Long updatedTime;
