+++ vire-model/src/main/java/com/vire/model/response/FirmResponse.java
@@ -24,7 +24,7 @@ public class FirmResponse {
     private Long thumbsUpCount = 0l;
     private Long thumbsDownCount = 0l;
     private Long friendsCount = 0l;
-    private Long starsCount = 0l;
+    private Integer starsCount = 0;
     private List<ProfileSettingResponse> profileSettingTypes;
     private FirmProfileResponse firmProfile;
